+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.4.0](https://www.github.com/orthanq/orthanq/compare/v1.3.0...v1.4.0) (2024-06-20)
+
+
+### Features
+
+* download nextstrain resources and generate candidate variants from nuc conversions ([#12](https://www.github.com/orthanq/orthanq/issues/12)) ([f1ffda9](https://www.github.com/orthanq/orthanq/commit/f1ffda92f328d4078746b0b374b9e9dc93f8037e))
+
+## [1.3.0](https://www.github.com/orthanq/orthanq/compare/v1.2.0...v1.3.0) (2024-04-12)
+
+
+### Features
+
+* breaking change: orthanq virus candidates: use the oldest submitted genome to represent each nextstrain in candidate haplotypes avoiding  irrelevant pango lineages ([96f397b](https://www.github.com/orthanq/orthanq/commit/96f397b0ba99f0f695b93b75303da3ed0d269b52))
+
 ## [1.2.0](https://www.github.com/orthanq/orthanq/compare/v1.1.0...v1.2.0) (2024-02-05)